backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
redis==5.0.3
python-dotenv==1.0.1
pydantic==2.6.4
qrcode[pil]==7.4.2
alembic==1.13.1 # Uncomment if using Alembic for migrations

# Testing dependencies
pytest==8.1.1
httpx==0.27.0
fakeredis[lua]==2.22.0